Set tanLambda range to -8 to +8

+++ L1Trigger/TrackFindingTMTT/python/TMTrackProducer_Defaults_cfi.py
@@ -462,7 +462,7 @@ TMTrackProducer_params = cms.PSet(
     MiniHT_z0Bits = cms.uint32(10),
     MiniHT_z0Range  = cms.double(60),
     MiniHT_tanlambdaBits = cms.uint32(14),
-    MiniHT_tanlambdaRange = cms.double(32),
+    MiniHT_tanlambdaRange = cms.double(16),
     
     #======= SimpleLR digi parameters ========
     SLR_skipTrackDigi = cms.bool( False ), # Optionally skip track digitisation if done internally inside fitting code.
@@ -490,7 +490,7 @@ TMTrackProducer_params = cms.PSet(
     KF_z0Bits = cms.uint32(12),
     KF_z0Range  = cms.double(60),
     KF_tanlambdaBits = cms.uint32(16),
-    KF_tanlambdaRange = cms.double(32),
+    KF_tanlambdaRange = cms.double(16),
     # KF_chisquaredBits = cms.uint32(4), // Number of bits when we move to binned chi2
     KF_chisquaredBits = cms.uint32(17),
     KF_chisquaredRange = cms.double(1024.),
